+++ Tests/URLTests/URLTests.swift
@@ -1,46 +1,25 @@
 import XCTest
 @testable import URL
 
+// TODO:
+// Test plan.
+//================
+// - PercentEscaping: long strings, short strings, escape sets, unescaping, round-tripping
+// - hasNonURLCodePoints
+// - WebURLParser.Components.QueryParameters
+
+let testBasic_printResults = true
+
 final class URLTests: XCTestCase {
 
-    func testSpecExamples() {
-        // These examples taken from the spec:
-        // https://url.spec.whatwg.org/#urls
-
-        if let results = WebURLParser.parse("https://example.org/") {
-            XCTAssertEqual(results.scheme, .https)
-            XCTAssertEqual(results.host, .domain("example.org"))
-        } else {
-            XCTFail("Failed to parse")
-        }
-
-        if let results = WebURLParser.parse("https://////example.org///") {
-            XCTAssertEqual(results.scheme, .https)
-            XCTAssertEqual(results.host, .domain("example.org"))
-        } else {
-            XCTFail("Failed to parse")
-        }
-	
-        if let results = WebURLParser.parse("https://example.com/././foo") {
-            XCTAssertEqual(results.scheme, .https)
-            XCTAssertEqual(results.host, .domain("example.com"))
-            XCTAssertEqual(results.path, ["foo"])
-        } else {
-            XCTFail("Failed to parse")
-        }
-
-         if let results = WebURLParser.parse("https://EXAMPLE.com/../x") {
-             XCTAssertEqual(results.scheme, .https)
-             XCTAssertEqual(results.host, .domain("example.com"))
-             XCTAssertEqual(results.path, ["x"])
-         } else {
-             XCTFail("Failed to parse")
-         }
-    }
-
+   /// Tests a handful of basic situations demonstrating the major features of the parser.
+   /// These tests are not meant to be exhaustive; for something more comprehensive, see the WHATWG constructor tests.
+   ///
+   /// Note that these tests operate at the `WebURLParser.Components` level, not the `WebURL` object model-level.
+   ///
    func testBasic() {
 
-       let testData: [(String, WebURLParser.Components)] = [
+       let testData: [(String, WebURLParser.Components?)] = [
 
         // Leading, trailing whitespace.
         ("        http://www.google.com   ", WebURLParser.Components(
@@ -50,9 +29,6 @@ final class URLTests: XCTestCase {
         ),
 
         // Non-ASCII characters in path.
-        //
-        // FIXME: Includes empty path components, but I'm *sure* this follows the spec
-        // (in fact, we deviate a little so we don't add quite so many empty path components).
         ("http://mail.yahoo.com/€uronews/", WebURLParser.Components(
             scheme: .http,
             username: "", password: "", host: .domain("mail.yahoo.com"), port: nil,
@@ -74,9 +50,6 @@ final class URLTests: XCTestCase {
         ),
 
         // '..' in path.
-        //
-        // FIXME: Includes empty path components, but I'm *sure* this follows the spec
-        // (in fact, we deviate a little so we don't add quite so many empty path components).
         ("http://www.test.com/../athing/anotherthing/.././something/", WebURLParser.Components(
             scheme: .http,
             username: "", password: "", host: .domain("www.test.com"), port: nil,
@@ -96,6 +69,16 @@ final class URLTests: XCTestCase {
             username: "", password: "", host: .ipv4Address(IPAddress.V4("192.168.0.1")!), port: nil,
             path: ["aThing"], query: nil, fragment: nil, cannotBeABaseURL: false)
         ),
+        
+        // Invalid IPv4 address (trailing non-hex-digit makes it a domain).
+        ("https://0x3h", WebURLParser.Components(
+            scheme: .https,
+            username: "", password: "", host: .domain("0x3h"), port: nil,
+            path: [""], query: nil, fragment: nil, cannotBeABaseURL: false)
+        ),
+    	
+        // Invalid IPv4 address (overflows, otherwise correctly-formatted).
+       ("https://234.266.2", nil),
 
         // Non-ASCII opaque host.
         ("tp://www.bücher.de", WebURLParser.Components(
@@ -115,71 +98,67 @@ final class URLTests: XCTestCase {
         // ==== Everything below is XFAIL while host parsing is still being implemented ==== //
 
         //Non-ASCII domain.
-        // ("http://www.bücher.de", XURL.Components(
+        // ("http://www.bücher.de", WebURLParser.Components(
         //     scheme: "http",
         //     authority: .init(username: nil, password: nil, host: .domain("www.xn--bcher-kva.de"), port: nil),
         //     path: [""], query: nil, fragment: nil, cannotBeABaseURL: false)
         // ),
        ]
+    
+        fileprivate func debugPrint(_ url: String, _ parsedComponents: WebURLParser.Components?) {
+            print("URL:\t|\(url)|")
+            if let results = parsedComponents {
+                print("Results:\n\(WebURL(components: results).debugDescription)")
+            } else {
+                print("Results:\nFAIL")
+            }
+            print("===================")
+        }
 
-       print("===================")
+    
+       if testBasic_printResults {
+           print("===================")
+       }
        for (input, expectedComponents) in testData {
             let results = WebURLParser.parse(input)
             XCTAssertEqual(results, expectedComponents, "Failed to correctly parse \(input)")
-            debugPrint(input, results)
+            if testBasic_printResults {
+                debugPrint(input, results)
+            }
         }
-        print("===================")
-   }
-
-   func testInvalidIPv4() {
-        let testData: [(String, WebURLParser.Components?)] = [
-            ("https://0x3h", WebURLParser.Components(
-                scheme: .https,
-                username: "", password: "", host: .domain("0x3h"), port: nil,
-                path: [""], query: nil, fragment: nil, cannotBeABaseURL: false)),
-            ("https://234.266.2", nil),
-        ]
-
-        for (input, expectedComponents) in testData {
-            let results = WebURLParser.parse(input)
-            XCTAssertEqual(results, expectedComponents, "Failed to correctly parse \(input)")
-        }
-   }
-
-   func testPercentEscaping() {
-       let testStrings: [String] = [
-         "hello, world", // ASCII
-         "👩‍👩‍👦‍👦️", // Long unicode
-         "%🐶️",   // Leading percent
-         "%z🐶️",  // Leading percent + one nonhex
-         "%3🐶️",  // Leading percent + one hex
-         "%3z🐶️", // Leading percent + one hex + one nonhex
-         "🐶️%",   // Trailing percent
-         "🐶️%z",  // Trailing percent + one nonhex
-         "🐶️%3",  // Trailing percent + one hex
-         "🐶️%3z", // Trailing percent + one hex + one nonhex
-         // "%100" FIXME: Percent escaping doesn't round-trip.
-       ]
-       for string in testStrings {
-//           let escaped = string.percentEscaped(where: { _ in false })
-//           let decoded = escaped.removingPercentEscaping() //PercentEscaping.decodeString(utf8: escaped.utf8)
-
-//           XCTAssertEqual(Array(string.utf8), Array(decoded.utf8))
-//            print("--------------")
-//            print("original: '\(string)'\t\tUTF8: \(Array(string.utf8))")
-//            print("decoded : '\(decoded)'\t\tUTF8: \(Array(decoded.utf8))")
-//            print("escaped:  '\(escaped)'\t\tUTF8: \(Array(escaped.utf8))")
+        if testBasic_printResults {
+            print("===================")
         }
    }
 }
 
+// Percent-escaping tests.
 
-fileprivate func debugPrint(_ url: String, _ parsedComponents: WebURLParser.Components?) {
-    print("URL:\t|\(url)|")
-    if let results = parsedComponents {
-        print("Results:\n\(results)")
-    } else {
-        print("Results:\nFAIL")
-    }
-    print("===================")
+extension URLTests {
+    
+     func testPercentEscaping() {
+           let testStrings: [String] = [
+             "hello, world", // ASCII
+             "👩‍👩‍👦‍👦️", // Long unicode
+             "%🐶️",   // Leading percent
+             "%z🐶️",  // Leading percent + one nonhex
+             "%3🐶️",  // Leading percent + one hex
+             "%3z🐶️", // Leading percent + one hex + one nonhex
+             "🐶️%",   // Trailing percent
+             "🐶️%z",  // Trailing percent + one nonhex
+             "🐶️%3",  // Trailing percent + one hex
+             "🐶️%3z", // Trailing percent + one hex + one nonhex
+             // "%100" FIXME: Percent escaping doesn't round-trip.
+           ]
+           for string in testStrings {
+    //           let escaped = string.percentEscaped(where: { _ in false })
+    //           let decoded = escaped.removingPercentEscaping() //PercentEscaping.decodeString(utf8: escaped.utf8)
+
+    //           XCTAssertEqual(Array(string.utf8), Array(decoded.utf8))
+    //            print("--------------")
+    //            print("original: '\(string)'\t\tUTF8: \(Array(string.utf8))")
+    //            print("decoded : '\(decoded)'\t\tUTF8: \(Array(decoded.utf8))")
+    //            print("escaped:  '\(escaped)'\t\tUTF8: \(Array(escaped.utf8))")
+            }
+       }
 }
